+++ pkg/k8s/k8s_pvc.go
@@ -1,14 +1,14 @@
 package k8s
 
 import (
-	"context"
-	"fmt"
-	"time"
+    "context"
+    "fmt"
+    "time"
 
-	"github.com/sirupsen/logrus"
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+    "github.com/sirupsen/logrus"
+    v1 "k8s.io/api/core/v1"
+    "k8s.io/apimachinery/pkg/api/resource"
+    metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 // createPersistentVolumeClaim deploys a PersistentVolumeClaim if it does not exist.
@@ -32,6 +32,9 @@ func createPersistentVolumeClaim(namespace, name string, labels map[string]strin
 	ctx, cancel := context.WithTimeout(context.Background(), 20*time.Second)
 	defer cancel()
 
+	if !IsInitialized() {
+		return fmt.Errorf("knuu is not initialized")
+	}
 	if _, err := Clientset().CoreV1().PersistentVolumeClaims(namespace).Create(ctx, pvc, metav1.CreateOptions{}); err != nil {
 		return err
 	}
@@ -52,6 +55,9 @@ func deletePersistentVolumeClaim(namespace, name string) error {
 	ctx, cancel := context.WithTimeout(context.Background(), 20*time.Second)
 	defer cancel()
 
+	if !IsInitialized() {
+		return fmt.Errorf("knuu is not initialized")
+	}
 	if err := Clientset().CoreV1().PersistentVolumeClaims(namespace).Delete(ctx, name, metav1.DeleteOptions{}); err != nil {
 		return fmt.Errorf("error deleting PersistentVolumeClaim %s: %w", name, err)
 	}
@@ -66,6 +72,9 @@ func getPersistentVolumeClaim(namespace, name string) (*v1.PersistentVolumeClaim
 	ctx, cancel := context.WithTimeout(context.Background(), 20*time.Second)
 	defer cancel()
 
+	if !IsInitialized() {
+		return nil, fmt.Errorf("knuu is not initialized")
+	}
 	pv, err := Clientset().CoreV1().PersistentVolumeClaims(namespace).Get(ctx, name, metav1.GetOptions{})
 	if err != nil {
 		return nil, err
